+++ lib/features/dashboard/domain/repo/analytics_repo.dart
@@ -2,7 +2,7 @@ import 'package:dartz/dartz.dart';
 import 'package:shop_sphere/core/errors/fairebase_failure.dart';
 
 abstract class AnalyticsRepo {
-  Future<Either<FirebaseFailure, int>> getOrdersTotalPriceTimeRange(
+  Future<Either<FirebaseFailure, double>> getOrdersTotalPriceTimeRange(
       {required int timeRangeIndex});
   Future<Either<FirebaseFailure, List<int>>> getDayOrdersTotalPrice();
 }
